slice example

+++ src/go08/dataType/dataType.go
@@ -3,7 +3,9 @@ package main
 import "fmt"
 
 func main() {
-	var a int = 12
-	var b int = 5
+	// var a int = 12
+	// var b int = 5
+	var a float64 = 12
+	var b float64 = 5
 	fmt.Println(a % b)
 }
